+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Most recent change on the bottom.
 ## [Unreleased] - 0.5.6
 ### Added
 - `nequip-benchmark` and `nequip-train` report number of weights and number of trainable weights
+- `nequip-benchmark --no-compile` and `--verbose` and `--memory-summary`
+- `nequip-benchmark --pdb` for debugging model (builder) errors
+
+### Fixed
+- Git utilities when installed as ZIPed `.egg` (#264)
 
 ## [0.5.5] - 2022-06-20
 ### Added